reduce stack usage Bench: 6712469

+++ src/search.h
@@ -10,10 +10,10 @@ struct SearchStack {
     int16_t staticEval;
     Move excludedMove;
     Move move;
-    int ply;
+    uint16_t ply;
     Move searchKiller;
     int (*contHistEntry)[12 * 64];
-    int reduction;
+    int16_t reduction;
 };
